feat: use per-frame 3D item rendering for inventory icons
