+++ pomme-client/src/renderer/util.rs
@@ -35,120 +35,6 @@ pub fn create_gpu_image_with_format(
     (image, view, allocation)
 }
 
-pub fn create_color_attachment_image(
-    device: &vk::Device,
-    allocator: &Arc<Mutex<Allocator>>,
-    width: u32,
-    height: u32,
-    format: vk::Format,
-    name: &str,
-) -> (vk::Image, vk::ImageView, Allocation) {
-    let usage = vk::ImageUsageFlags::ColorAttachment
-        | vk::ImageUsageFlags::Sampled
-        | vk::ImageUsageFlags::TransferDst;
-    create_attachment_image(
-        device,
-        allocator,
-        width,
-        height,
-        format,
-        usage,
-        vk::ImageAspectFlags::Color,
-        name,
-    )
-}
-
-pub fn create_depth_attachment_image(
-    device: &vk::Device,
-    allocator: &Arc<Mutex<Allocator>>,
-    width: u32,
-    height: u32,
-    format: vk::Format,
-    name: &str,
-) -> (vk::Image, vk::ImageView, Allocation) {
-    let usage = vk::ImageUsageFlags::DepthStencilAttachment;
-    create_attachment_image(
-        device,
-        allocator,
-        width,
-        height,
-        format,
-        usage,
-        vk::ImageAspectFlags::Depth,
-        name,
-    )
-}
-
-#[allow(clippy::too_many_arguments)]
-fn create_attachment_image(
-    device: &vk::Device,
-    allocator: &Arc<Mutex<Allocator>>,
-    width: u32,
-    height: u32,
-    format: vk::Format,
-    usage: vk::ImageUsageFlags,
-    aspect_mask: vk::ImageAspectFlags,
-    name: &str,
-) -> (vk::Image, vk::ImageView, Allocation) {
-    let image_info = vk::ImageCreateInfo {
-        image_type: vk::ImageType::Type2D,
-        format,
-        extent: vk::Extent3D {
-            width,
-            height,
-            depth: 1,
-        },
-        mip_levels: 1,
-        array_layers: 1,
-        samples: vk::SampleCountFlags::Type1,
-        tiling: vk::ImageTiling::Optimal,
-        usage,
-        ..Default::default()
-    };
-
-    let image = device
-        .create_image(&image_info, None)
-        .expect("failed to create attachment image");
-    let mem_reqs = device.get_image_memory_requirements(image);
-
-    let allocation = allocator
-        .lock()
-        .unwrap()
-        .allocate(&AllocationCreateDesc {
-            name,
-            requirements: mem_reqs,
-            location: MemoryLocation::GpuOnly,
-            linear: false,
-            allocation_scheme: AllocationScheme::GpuAllocatorManaged,
-        })
-        .expect("failed to allocate attachment image memory");
-
-    unsafe {
-        device
-            .bind_image_memory(image, allocation.memory(), allocation.offset())
-            .expect("failed to bind attachment image memory");
-    }
-
-    let view_info = vk::ImageViewCreateInfo {
-        image,
-        view_type: vk::ImageViewType::Type2D,
-        format,
-        subresource_range: vk::ImageSubresourceRange {
-            aspect_mask,
-            base_mip_level: 0,
-            level_count: 1,
-            base_array_layer: 0,
-            layer_count: 1,
-        },
-        ..Default::default()
-    };
-    let view = device
-        .create_image_view(&view_info, None)
-        .expect("failed to create attachment image view");
-
-    (image, view, allocation)
-}
-
 fn create_gpu_image_core(
     device: &vk::Device,
     allocator: &Arc<Mutex<Allocator>>,
